#89 [refactor] : 클래스명을 변경한다
